move all resources into the resources dir
