+++ pkg/infra/iac2/templates_compiler_test.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/klothoplatform/klotho/pkg/core"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/s3"
+	"github.com/klothoplatform/klotho/pkg/provider/aws/resources"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -167,7 +167,7 @@ func Test_handleIaCValue(t *testing.T) {
 		{
 			name: "bucket name",
 			value: core.IaCValue{
-				Resource: s3.NewS3Bucket(&core.Fs{}, "test-app"),
+				Resource: resources.NewS3Bucket(&core.Fs{}, "test-app"),
 				Property: string(core.BUCKET_NAME),
 			},
 			resourceVarNamesById: map[string]string{
