+++ pkg/provider/aws/execution_unit.go
@@ -6,10 +6,6 @@ import (
 	"github.com/klothoplatform/klotho/pkg/config"
 	"github.com/klothoplatform/klotho/pkg/core"
 	"github.com/klothoplatform/klotho/pkg/provider/aws/resources"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/cloudwatch"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/ecr"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/iam"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/lambda"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
@@ -20,12 +16,12 @@ func (a *AWS) GenerateExecUnitResources(unit *core.ExecutionUnit, result *core.C
 
 	execUnitCfg := a.Config.GetExecutionUnit(unit.ID)
 
-	image, err := ecr.GenerateEcrRepoAndImage(a.Config.AppName, unit, dag)
+	image, err := resources.GenerateEcrRepoAndImage(a.Config.AppName, unit, dag)
 	if err != nil {
 		return err
 	}
 
-	role := iam.NewIamRole(a.Config.AppName, fmt.Sprintf("%s-ExecutionRole", unit.ID), unit.Provenance(), GetAssumeRolePolicyForType(execUnitCfg))
+	role := resources.NewIamRole(a.Config.AppName, fmt.Sprintf("%s-ExecutionRole", unit.ID), unit.Provenance(), GetAssumeRolePolicyForType(execUnitCfg))
 	dag.AddResource(role)
 	err = a.PolicyGenerator.AddUnitRole(unit.Id(), role)
 	if err != nil {
@@ -43,9 +39,9 @@ func (a *AWS) GenerateExecUnitResources(unit *core.ExecutionUnit, result *core.C
 	switch execUnitCfg.Type {
 	case Lambda:
 
-		lambdaFunction := lambda.NewLambdaFunction(unit, a.Config.AppName, role, image)
+		lambdaFunction := resources.NewLambdaFunction(unit, a.Config.AppName, role, image)
 		a.ConstructIdToResourceId[unit.Id()] = lambdaFunction.Id()
-		logGroup := cloudwatch.NewLogGroup(a.Config.AppName, fmt.Sprintf("/aws/lambda/%s", lambdaFunction.Name), unit.Provenance(), 5)
+		logGroup := resources.NewLogGroup(a.Config.AppName, fmt.Sprintf("/aws/lambda/%s", lambdaFunction.Name), unit.Provenance(), 5)
 		dag.AddResource(lambdaFunction)
 		dag.AddResource(logGroup)
 		dag.AddDependency(logGroup, lambdaFunction)
@@ -106,7 +102,7 @@ func (a *AWS) convertExecUnitParams(result *core.ConstructGraph, dag *core.Resou
 				return fmt.Errorf("resource not found for id, %s", resourceId)
 			}
 			switch r := resource.(type) {
-			case *lambda.LambdaFunction:
+			case *resources.LambdaFunction:
 				r.EnvironmentVariables = resourceEnvVars
 			}
 		} else {
@@ -120,15 +116,15 @@ func (a *AWS) convertExecUnitParams(result *core.ConstructGraph, dag *core.Resou
 func GetAssumeRolePolicyForType(cfg config.ExecutionUnit) string {
 	switch cfg.Type {
 	case Lambda:
-		return iam.LAMBDA_ASSUMER_ROLE_POLICY
+		return resources.LAMBDA_ASSUMER_ROLE_POLICY
 	case Ecs:
-		return iam.ECS_ASSUMER_ROLE_POLICY
+		return resources.ECS_ASSUMER_ROLE_POLICY
 	case Eks:
 		eksConfig := cfg.GetExecutionUnitParamsAsKubernetes()
 		if eksConfig.NodeType == string(resources.Fargate) {
-			return iam.EKS_FARGATE_ASSUME_ROLE_POLICY
+			return resources.EKS_FARGATE_ASSUME_ROLE_POLICY
 		}
-		return iam.EC2_ASSUMER_ROLE_POLICY
+		return resources.EC2_ASSUMER_ROLE_POLICY
 	}
 	return ""
 }
