+++ pkg/provider/aws/provider.go
@@ -2,11 +2,11 @@ package aws
 
 import (
 	"github.com/klothoplatform/klotho/pkg/config"
-	"github.com/klothoplatform/klotho/pkg/provider/aws/resources/iam"
+	"github.com/klothoplatform/klotho/pkg/provider/aws/resources"
 )
 
 type AWS struct {
 	Config                  *config.Application
 	ConstructIdToResourceId map[string]string
-	PolicyGenerator         *iam.PolicyGenerator
+	PolicyGenerator         *resources.PolicyGenerator
 }
