--- pkg/provider/aws/resources/ecr/ecr.go
@@ -1,30 +1,0 @@
-package ecr
-
-import (
-	"fmt"
-
-	"github.com/klothoplatform/klotho/pkg/core"
-)
-
-func GenerateEcrRepoAndImage(appName string, unit *core.ExecutionUnit, dag *core.ResourceGraph) (*EcrImage, error) {
-	// See if we have already created an ecr repository for the app and if not create one, otherwise add a ref to this exec unit
-	var repo *EcrRepository
-	existingRepo := dag.GetResource(GenerateRepoId(appName))
-	if existingRepo == nil {
-		repo = NewEcrRepository(appName, unit.Provenance())
-		dag.AddResource(repo)
-	} else {
-		var ok bool
-		repo, ok = existingRepo.(*EcrRepository)
-		if !ok {
-			return nil, fmt.Errorf("expected resource with id, %s, to be ecr repository", repo.Id())
-		}
-		repo.ConstructsRef = append(repo.ConstructsRef, unit.Provenance())
-	}
-
-	// Create image and make it dependent on the repository
-	image := NewEcrImage(unit, appName, repo)
-	dag.AddResource(image)
-	dag.AddDependency(repo, image)
-	return image, nil
-}
