+++ pkg/provider/aws/resources/ecr_test.go
@@ -1,4 +1,4 @@
-package ecr
+package resources
 
 import (
 	"testing"
